Split task lists into 'master' and 'displayed'
The master task list will contain all the tasks in memory.
The displayed task list will be what is displayed to the user.

This change is in preparation for search functionality, where the
displayed tasks will have to be modified yet the full (master) list of
tasks still need to be kept in memory.