Refactor parsing structure
The logic responsible to creating the relevant Command object based on
the user supplied input string has been moved to the CommandController
from the CommandParser. This is to separate the parsing logic from any
other logic dealing with manipulating commands.

A basic skeleton for the info view has also been added.