Add actions to be executed after view is updated
Some commands require additional actions to be executed on the view for
usability reasons. For instance, after adding a task, the task list is
scrolled to the recently added task. In this example, an action is
registered with the ViewState. This action contains an ActionType
(SCROLL_TASK_LIST_TO) and an associated Object (task).

These registered actions are executed after the view is updated.